Feature: 댓글, 태그 기능 API 구현 (#5)

+++ src/main/java/com/traffic/practice/mapper/PostMapper.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Mapper
 public interface PostMapper {
 
-    public int register(PostDTO postDTO);
+    public Long register(PostDTO postDTO);
     public List<PostDTO> selectMyPostList(Long userId);
     public void updatePost(PostDTO postDTO);
     public void deletePost(Long userId, Long id);
